+++ packages/blocks-core/php/libs/wordpress/group/block.php
@@ -25,7 +25,27 @@ return array_merge(
 				blockera_load( 'inners.button', dirname( __DIR__ ) ),
 				blockera_load( 'inners.headings', dirname( __DIR__ ) ),
 				blockera_load( 'inners.paragraph', dirname( __DIR__ ) ),
-			)
+			),
+			[
+				'blockera/elements/bold' => [
+					'root' => 'strong, b',
+				],
+				'blockera/elements/italic' => [
+					'root' => 'em, i',
+				],
+				'blockera/elements/kbd' => [
+					'root' => 'kbd',
+				],
+				'blockera/elements/code' => [
+					'root' => 'code',
+				],
+				'blockera/elements/span' => [
+					'root' => 'span',
+				],
+				'blockera/elements/mark' => [
+					'root' => 'mark',
+				],
+			]
 		),
 	]
 );