+++ packages/blocks-core/CHANGELOG.md
@@ -30,6 +30,7 @@
 - `Comment Content` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Post Excerpt` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Details` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Group` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
 ### Automated Tests
@@ -61,6 +62,7 @@
 - Added E2E tests to check functionality `Comment Content` new inner block.
 - Added E2E tests to check functionality `Post Excerpt` new inner block.
 - Added E2E tests to check functionality `Details` new inner block.
+- Added E2E tests to check functionality `Group` new inner block.
 
 ## 1.4.0 (2025-05-19)